ghci: fix host/target word size mismatch in BreakArray logic
This patch fixes host/target word size mismatch in BreakArray logic by
specifying each element to be an Int64 unconditionally, instead of a
machine word. For cross GHCs with 64-bit host GHC and 32-bit target
(e.g. wasm32), the ghci library is compiled and executed on both
host/target, and the mismatch here would cause the ghci breakpoint
functionality to arbitrarily break (e.g. #25420).

+++ compiler/GHC/Runtime/Eval.hs
@@ -135,6 +135,7 @@ import Control.Monad
 import Control.Monad.Catch as MC
 import Data.Array
 import Data.Dynamic
+import Data.Int
 import Data.IntMap (IntMap)
 import qualified Data.IntMap as IntMap
 import Data.List (find,intercalate)
@@ -400,7 +401,7 @@ handleRunStatus step expr bindings final_ids status history
     = return (ExecComplete (Left (fromSerializableException e)) alloc)
 
 
-resumeExec :: GhcMonad m => (SrcSpan->Bool) -> SingleStep -> Maybe Int
+resumeExec :: GhcMonad m => (SrcSpan->Bool) -> SingleStep -> Maybe Int64
            -> m ExecResult
 resumeExec canLogSpan step mbCnt
  = do
@@ -458,7 +459,7 @@ resumeExec canLogSpan step mbCnt
                                                         fromListBL 50 hist
                 handleRunStatus step expr bindings final_ids status hist'
 
-setupBreakpoint :: GhcMonad m => HscEnv -> BreakpointId -> Int -> m ()   -- #19157
+setupBreakpoint :: GhcMonad m => HscEnv -> BreakpointId -> Int64 -> m ()   -- #19157
 setupBreakpoint hsc_env bi cnt = do
   let modl = bi_tick_mod bi
       breaks hsc_env modl = getModBreaks $ expectJust "setupBreakpoint" $
